Fixed a bug with our budget summary endpoint where a byte was being returned instead of a valid numerial value causing a 500 error response.

+++ internal/sql/queries/financial_manager_queries.sql
@@ -355,10 +355,10 @@ SELECT
     COALESCE(res.recurring_expenses, '[]'::jsonb) AS recurring_expenses,
 
     -- Total projected recurring expenses for each budget
-    COALESCE(res.total_projected_recurring_expenses, 0) AS total_projected_recurring_expenses,
+    COALESCE(res.total_projected_recurring_expenses, 0)::NUMERIC AS total_projected_recurring_expenses,
 
     -- Total non-recurring expenses for each budget
-    COALESCE(es.total_expenses, 0) AS total_expenses
+    COALESCE(es.total_expenses, 0)::NUMERIC AS total_expenses
     
 FROM budgets b
 LEFT JOIN goals g ON b.id = g.budget_id
